DCT's new DFTensor
