+++ psi4/src/psi4/dct/dct_integrals_RHF.cc
@@ -48,7 +48,7 @@ void DCTSolver::transform_integrals_RHF() {
 
     if (options_.get_str("DCT_TYPE") == "DF") {
         // Transform b(Q|mn) to b(Q|pq) in MO basis
-        transform_b();
+        transform_b_so2mo();
 
         psio_->open(PSIF_LIBTRANS_DPD, PSIO_OPEN_OLD);
