+++ psi4/src/psi4/dct/dct_integrals_UHF.cc
@@ -74,7 +74,7 @@ void DCTSolver::transform_integrals() {
     _ints->update_orbitals();
     if (options_.get_str("DCT_TYPE") == "DF") {
         // Transform b(Q|mn) to b(Q|pq) in MO basis
-        transform_b();
+        transform_b_so2mo();
         psio_->open(PSIF_LIBTRANS_DPD, PSIO_OPEN_OLD);
         /*- Transform g(OV|OV) -*/
         form_df_g_ovov();
